+++ Sources/CodeViewerKitLanguageRuby/RubyGrammar.swift
@@ -1,7 +1,0 @@
-import CodeViewerKit
-import TreeSitterRuby
-import TreeSitterRubyQueries
-
-public extension CodeGrammar {
-    static let ruby = CodeGrammar(identifier: "ruby", aliases: ["rb"], language: tree_sitter_ruby(), queryURLs: [TreeSitterRubyQueries.Query.highlightsFileURL])
-}